refactor(web): co-locate foldered components' tests in their own __tests__
Move Sidebar, RequestForm, and ExampleAccordion tests into
components/<Name>/__tests__/ beside each component. Single-file components
keep sharing components/__tests__/. Document the rule in AGENTS.md.
