+++ web/src/components/ExampleAccordion/__tests__/ExampleAccordion.test.tsx
@@ -1,8 +1,8 @@
 // @vitest-environment jsdom
 import { describe, it, expect, afterEach, beforeEach } from "vitest";
 import { render, screen, fireEvent, cleanup } from "@testing-library/react";
-import { ExampleAccordion } from "../ExampleAccordion/ExampleAccordion";
-import type { Example } from "../../api";
+import { ExampleAccordion } from "../ExampleAccordion";
+import type { Example } from "../../../api";
 
 const example: Example = {
   name: "returns the user",
