+++ web/src/components/Sidebar/__tests__/Sidebar.test.tsx
@@ -1,8 +1,8 @@
 // @vitest-environment jsdom
 import { describe, it, expect, vi, afterEach } from "vitest";
 import { render, screen, fireEvent, cleanup } from "@testing-library/react";
-import { Sidebar } from "../Sidebar/Sidebar";
-import type { Endpoint } from "../../api";
+import { Sidebar } from "../Sidebar";
+import type { Endpoint } from "../../../api";
 
 const endpoints: Endpoint[] = [
   {
